requirements.txt
Django>=1.3
django-admin-tools
coverage
django-coverage
tddspry
django-request
-e hg+https://bitbucket.org/vice_versa/django-simple-history#egg=simple_history
django-webtest
webtest
pillow
sorl-thumbnail
PIL
django-positions
south
